Replaces uses of doubles with floats
Looks like the KallistiOS Dreamcast SDK disables double support
by default: https://github.com/KallistiOS/KallistiOS/blob/495e77fd60d5b09a1ad52a26cd4a7e73cd0d9d51/environ_dreamcast.sh#L16

We don't really need doubles in this code.

The one place where we might have needed them is the SMK video
decoder, handled in a separate PR.

+++ Source/monster.cpp
@@ -832,7 +832,7 @@ void DiabloDeath(Monster &diablo, bool sendmsg)
 		dist = 20;
 	diablo.var3 = ViewPosition.x << 16;
 	diablo.position.temp.x = ViewPosition.y << 16;
-	diablo.position.temp.y = (int)((diablo.var3 - (diablo.position.tile.x << 16)) / (double)dist);
+	diablo.position.temp.y = (int)((diablo.var3 - (diablo.position.tile.x << 16)) / (float)dist);
 	if (!gbIsMultiplayer) {
 		Player &myPlayer = *MyPlayer;
 		myPlayer.pDiabloKillLevel = std::max(myPlayer.pDiabloKillLevel, static_cast<uint8_t>(sgGameInitInfo.nDifficulty + 1));
